tag filter: refactor of PR #10521
